Update payment component layout and add library reference
Modified the layout of buttons in `MakePaymentComponent.razor`, repositioning existing buttons and adding a new `LabelGridComponent`. Added a package reference for `CommonBasicLibraries` version `9.0.17` in `DealCardGame.csproj` to enhance project capabilities.  Did this to prevent players from accidently choosing to choose all cards for payments when you intended to only select some for payment.